requirements.txt
Django==2.2.8
djangorestframework==3.9.4
requests==2.22.0